+++ src/test/java/com/yandex/diplom2/Steps/BaseSteps.java
@@ -1,0 +1,10 @@
+package com.yandex.diplom2.Steps;
+
+import io.restassured.response.Response;
+
+public class BaseSteps {
+    protected static Response responseOfUserCreation;
+    protected static Response responseOfUserLogin;
+    protected static Response responseOfUserNotRegistrationAndLogin;
+    protected static String json;
+}
